+++ src/chat/chat-room/proxy-chat-room.h
@@ -80,7 +80,7 @@ public:
 
 	void markAsRead () override;
 	void enableEphemeral (bool ephem) override;
-	void setEphemeralLifetime (double lifetime) override;
+	void setEphemeralLifetime (long lifetime) override;
 
 	const IdentityAddress &getConferenceAddress () const override;
